weight 더블형으로 변경

+++ app/src/main/java/com/devlog/article/data/response/ArticleResponse.kt
@@ -18,7 +18,7 @@ data class Article(
     val title: String,
     val cx: Int,
     val _id : String,
-    var weight : Int?
+    var weight : Double?
 ): Serializable  {
 
 }
